[refactor] remove warnning
- change tmpfile() to tmpfile_s()
- change type from size_t to int

+++ TestShell/RealFileIo.cpp
@@ -20,7 +20,7 @@ size_t RealFileIo::Read(int fd, void* buf, size_t count)
         throw runtime_error("failed to load file");
     }
 
-    fgets((char*)buf, count, file);
+    fgets((char*)buf, (int)count, file);
     fclose(file);
 
     return count;
